requirements.txt
discord.py
requests
regex